requirements.txt
python-decouple==3.8
invoke==2.2.0
streamlit==1.28.2
colouredlogs==10.0.1
sh==2.0.6
beautifulsoup4==4.12.2
requests-cache==1.1.1
